requirements/dev.txt
-r base.txt  # includes travis.txt requirements file

colorama==0.3.3
django-coverage==1.2
django-debug-toolbar==1.4
django-extensions==1.5
fake-factory==0.5.3
ipdb==0.8.1

# tests
astroid==1.4.8
codecov==1.4.1
coverage==4.0
django-nose==1.4.2
pep8==1.6.2
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
